added new frontend
