+++ Frontend/src/App.css
@@ -1,28 +1,161 @@
-.App {
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600;700&display=swap');
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  font-family: 'Poppins', sans-serif;
+}
+
+.app {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  transition: background 0.3s ease, color 0.3s ease;
+}
+
+.app.dark {
+  background: linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%);
+  color: #fff;
+}
+
+.app.light {
+  background: linear-gradient(135deg, #fefefe 0%, #f0f4f8 100%);
+  color: #2c3e50;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px 40px;
+  background: rgba(255, 255, 255, 0.05);
+  backdrop-filter: blur(10px);
+  box-shadow: 0 2px 15px rgba(0, 0, 0, 0.05);
+}
+
+.app.light .header {
+  background: rgba(255, 255, 255, 0.9);
+}
+
+.logo {
+  font-size: 1.8rem;
+  font-weight: 700;
+  letter-spacing: 1px;
+  text-transform: uppercase;
+}
+
+.app.dark .logo {
+  color: #fff;
+}
+
+.app.light .logo {
+  color: #2c3e50;
+}
+
+nav {
+  display: flex;
+  gap: 15px;
+}
+
+nav button {
+  padding: 12px 25px;
+  border: none;
+  border-radius: 25px;
+  font-size: 1rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: background 0.3s ease, transform 0.2s ease;
+}
+
+.theme-toggle {
+  background: #f1c40f;
+  color: #2c3e50;
+}
+
+.theme-toggle:hover {
+  background: #f39c12;
+  transform: scale(1.05);
+}
+
+.app.dark nav button:not(.theme-toggle) {
+  background: #00d4ff;
+  color: #1a1a1a;
+}
+
+.app.dark nav button:not(.theme-toggle):hover {
+  background: #66e6ff;
+  transform: scale(1.05);
+}
+
+.app.light nav button:not(.theme-toggle) {
+  background: #1abc9c;
+  color: #fff;
+}
+
+.app.light nav button:not(.theme-toggle):hover {
+  background: #16a085;
+  transform: scale(1.05);
+}
+
+main {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+}
+
+.welcome {
+  text-align: center;
+  padding: 60px 20px;
+  animation: fadeIn 1s ease;
+}
+
+.app.dark .welcome h1 {
+  font-size: 3rem;
+  font-weight: 700;
+  margin-bottom: 20px;
+  text-shadow: 0 2px 5px rgba(0, 0, 0, 0.5);
+  color: #fff;
+}
+
+.app.light .welcome h1 {
+  font-size: 3rem;
+  font-weight: 700;
+  margin-bottom: 20px;
+  text-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  color: #2c3e50;
+}
+
+.app.dark .welcome p {
+  font-size: 1.4rem;
+  margin-bottom: 30px;
+  color: #d0d0d0;
+}
+
+.app.light .welcome p {
+  font-size: 1.4rem;
+  margin-bottom: 30px;
+  color: #7f8c8d;
+}
+
+.footer {
   text-align: center;
   padding: 20px;
+  font-size: 0.9rem;
+  width: 100%;
 }
 
-h1 {
-  color: #333;
+.app.dark .footer {
+  background: rgba(255, 255, 255, 0.03);
+  color: #d0d0d0;
 }
 
-form {
-  margin: 20px 0;
+.app.light .footer {
+  background: rgba(0, 0, 0, 0.02);
+  color: #7f8c8d;
 }
 
-input, button {
-  margin: 0 10px;
-  padding: 8px;
-}
-
-button {
-  background-color: #4CAF50;
-  color: white;
-  border: none;
-  cursor: pointer;
-}
-
-button:hover {
-  background-color: #45a049;
+@keyframes fadeIn {
+  from { opacity: 0; }
+  to { opacity: 1; }
 }
