+++ Frontend/src/components/Dashboard/Dashboard.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState, useMemo } from "react";
 import { Pie, Bar, Line, Doughnut } from "react-chartjs-2";
+import "../Dashboard/Dashboard.css";
 import {
   Chart as ChartJS,
   CategoryScale,
@@ -327,276 +328,5 @@ const Dashboard = ({ fileId }) => {
     </div>
   );
 };
-
-const styles = `
-  @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600;700&display=swap');
-
-  * {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    font-family: 'Poppins', sans-serif;
-  }
-
-  .dashboard {
-    padding: 40px 20px;
-    background: linear-gradient(135deg, #1e3c72 0%, #2a5298 100%);
-    min-height: 100vh;
-    color: #fff;
-    overflow-x: hidden;
-  }
-
-  .dashboard h2 {
-    text-align: center;
-    font-size: 2.8rem;
-    font-weight: 700;
-    color: #fff;
-    margin-bottom: 40px;
-    text-transform: uppercase;
-    letter-spacing: 2px;
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-    animation: fadeInDown 1s ease;
-  }
-
-  .status {
-    text-align: center;
-    padding: 20px;
-    border-radius: 12px;
-    font-size: 1.2rem;
-    margin: 0 auto 30px;
-    max-width: 600px;
-    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
-    transition: transform 0.3s ease;
-  }
-
-  .status.loading {
-    background: rgba(255, 255, 255, 0.1);
-    color: #00d4ff;
-    border: 2px solid #00d4ff;
-  }
-
-  .status.error {
-    background: rgba(255, 75, 75, 0.1);
-    color: #ff4b4b;
-    border: 2px solid #ff4b4b;
-  }
-
-  .summary, .key-takeaways, .trends {
-    background: #fff;
-    border-radius: 16px;
-    padding: 25px;
-    margin-bottom: 40px;
-    box-shadow: 0 8px 25px rgba(0, 0, 0, 0.15);
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-    animation: fadeInUp 0.8s ease;
-  }
-
-  .summary:hover, .key-takeaways:hover, .trends:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 12px 35px rgba(0, 0, 0, 0.2);
-  }
-
-  .summary h3, .key-takeaways h3, .trends h3 {
-    color: #1e3c72;
-    font-size: 2rem;
-    font-weight: 600;
-    margin-bottom: 20px;
-    border-bottom: 3px solid #2a5298;
-    display: inline-block;
-    padding-bottom: 5px;
-  }
-
-  .summary p {
-    font-size: 1.2rem;
-    color: #555;
-    margin: 10px 0;
-  }
-
-  .summary table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-top: 20px;
-    font-size: 1rem;
-    color: #333;
-  }
-
-  .summary th, .summary td {
-    padding: 15px;
-    border: 1px solid #e0e7ff;
-    text-align: left;
-  }
-
-  .summary th {
-    background: #2a5298;
-    color: #fff;
-    font-weight: 600;
-  }
-
-  .summary td {
-    background: #f9faff;
-  }
-
-  .key-takeaways ul {
-    list-style: none;
-    padding: 0;
-  }
-
-  .key-takeaways li {
-    color: #1e3c72;
-    font-size: 1.2rem;
-    margin-bottom: 15px;
-    padding-left: 30px;
-    position: relative;
-    transition: color 0.3s ease;
-  }
-
-  .key-takeaways li:hover {
-    color: #00d4ff;
-  }
-
-  .key-takeaways li:before {
-    content: "✔";
-    color: #27ae60;
-    position: absolute;
-    left: 0;
-    font-size: 1.4rem;
-    top: 2px;
-  }
-
-  .trends ul {
-    list-style: none;
-    padding: 0;
-  }
-
-  .trends li {
-    color: #555;
-    font-size: 1.2rem;
-    margin-bottom: 15px;
-    padding-left: 25px;
-    position: relative;
-  }
-
-  .trends li:before {
-    content: "➤";
-    color: #ff6384;
-    position: absolute;
-    left: 0;
-    font-size: 1.2rem;
-  }
-
-  .visualizations {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
-    gap: 30px;
-    padding: 0 10px;
-  }
-
-  .chart-container {
-    background: #fff;
-    border-radius: 16px;
-    padding: 25px;
-    box-shadow: 0 8px 25px rgba(0, 0, 0, 0.15);
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-    animation: fadeInUp 1s ease;
-  }
-
-  .chart-container:hover {
-    transform: translateY(-8px);
-    box-shadow: 0 12px 35px rgba(0, 0, 0, 0.25);
-  }
-
-  .chart-container h3 {
-    color: #1e3c72;
-    font-size: 1.6rem;
-    font-weight: 600;
-    margin-bottom: 20px;
-    text-align: center;
-  }
-
-  .chart-container p {
-    font-size: 1rem;
-    color: #777;
-    text-align: center;
-  }
-
-  .pagination {
-    margin-top: 40px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 20px;
-    flex-wrap: wrap;
-  }
-
-  .pagination button {
-    padding: 12px 25px;
-    background: #2a5298;
-    color: #fff;
-    border: none;
-    border-radius: 30px;
-    font-size: 1.1rem;
-    cursor: pointer;
-    transition: background 0.3s ease, transform 0.3s ease;
-  }
-
-  .pagination button:hover:not(:disabled) {
-    background: #00d4ff;
-    transform: scale(1.05);
-  }
-
-  .pagination button:disabled {
-    background: #b0c4de;
-    cursor: not-allowed;
-  }
-
-  .pagination span {
-    font-size: 1.2rem;
-    color: #fff;
-    font-weight: 600;
-  }
-
-  .pagination select {
-    padding: 10px;
-    border-radius: 8px;
-    border: none;
-    background: #fff;
-    color: #1e3c72;
-    font-size: 1.1rem;
-    cursor: pointer;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-  }
-
-  @keyframes fadeInDown {
-    from { opacity: 0; transform: translateY(-20px); }
-    to { opacity: 1; transform: translateY(0); }
-  }
-
-  @keyframes fadeInUp {
-    from { opacity: 0; transform: translateY(20px); }
-    to { opacity: 1; transform: translateY(0); }
-  }
-
-  @media (max-width: 768px) {
-    .dashboard h2 {
-      font-size: 2rem;
-    }
-    .summary h3, .key-takeaways h3, .trends h3 {
-      font-size: 1.6rem;
-    }
-    .summary table, .key-takeaways li, .trends li {
-      font-size: 1rem;
-    }
-    .chart-container h3 {
-      font-size: 1.4rem;
-    }
-    .visualizations {
-      grid-template-columns: 1fr;
-    }
-  }
-`;
-
-const styleSheet = document.createElement("style");
-styleSheet.textContent = styles;
-document.head.appendChild(styleSheet);
 
 export default Dashboard;
