+++ Frontend/src/components/FileUpload/FileUpload.css
@@ -1,0 +1,127 @@
+.file-upload-container {
+  padding: 25px;
+  border-radius: 12px;
+  max-width: 450px;
+  margin: 0 auto;
+  transition: all 0.3s ease;
+}
+
+.app.dark .file-upload-container {
+  background: rgba(255, 255, 255, 0.05);
+  box-shadow: 0 5px 20px rgba(0, 0, 0, 0.5);
+}
+
+.app.light .file-upload-container {
+  background: #fff;
+  box-shadow: 0 5px 25px rgba(0, 0, 0, 0.08);
+  border: 1px solid #ecf0f1;
+}
+
+.app.dark #get-started-btn {
+  padding: 15px 50px;
+  background: linear-gradient(90deg, #00d4ff, #66e6ff);
+  color: #1a1a1a;
+  border: none;
+  border-radius: 30px;
+  font-size: 1.3rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
+}
+
+.app.dark #get-started-btn:hover {
+  background: linear-gradient(90deg, #66e6ff, #00d4ff);
+  transform: scale(1.08);
+  box-shadow: 0 6px 25px rgba(0, 212, 255, 0.5);
+}
+
+.app.light #get-started-btn {
+  padding: 15px 50px;
+  background: linear-gradient(90deg, #1abc9c, #16a085);
+  color: #fff;
+  border: none;
+  border-radius: 30px;
+  font-size: 1.3rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
+}
+
+.app.light #get-started-btn:hover {
+  background: linear-gradient(90deg, #16a085, #1abc9c);
+  transform: scale(1.08);
+  box-shadow: 0 6px 25px rgba(26, 188, 156, 0.4);
+}
+
+.progress-bar {
+  height: 30px;
+  border-radius: 15px;
+  overflow: hidden;
+  margin-top: 20px;
+}
+
+.app.dark .progress-bar {
+  background: rgba(255, 255, 255, 0.1);
+  box-shadow: inset 0 2px 5px rgba(0, 0, 0, 0.2);
+}
+
+.app.light .progress-bar {
+  background: #ecf0f1;
+  box-shadow: inset 0 2px 5px rgba(0, 0, 0, 0.05);
+}
+
+.progress-fill {
+  height: 100%;
+  text-align: center;
+  line-height: 30px;
+  font-weight: 600;
+  transition: width 0.4s ease;
+}
+
+.app.dark .progress-fill {
+  background: #00d4ff;
+  color: #1a1a1a;
+}
+
+.app.light .progress-fill {
+  background: #1abc9c;
+  color: #fff;
+}
+
+.error-message {
+  text-align: center;
+  margin-top: 15px;
+  font-size: 1rem;
+  padding: 10px;
+  border-radius: 8px;
+}
+
+.app.dark .error-message {
+  color: #ff4b4b;
+  background: rgba(255, 75, 75, 0.1);
+}
+
+.app.light .error-message {
+  color: #e74c3c;
+  background: rgba(231, 76, 60, 0.1);
+}
+
+.success-message {
+  text-align: center;
+  margin-top: 15px;
+  font-size: 1rem;
+  padding: 10px;
+  border-radius: 8px;
+}
+
+.app.dark .success-message {
+  color: #27ae60;
+  background: rgba(39, 174, 96, 0.1);
+}
+
+.app.light .success-message {
+  color: #2ecc71;
+  background: rgba(46, 204, 113, 0.1);
+}
